Proper Support_Updater interfacing

+++ common/GlassUpdaterSupport.cs
@@ -194,6 +194,11 @@ function GlassUpdaterSupport::removeFromQueue(%obj) {
     messageBoxOk("Uh oh", "Updates for Blockland Glass are mandatory. It'll only take a minute!\n\nSorry about that.");
     return;
   }
+  if(%obj.name $= "Support_Updater") {
+    //must update!
+    messageBoxOk("Uh oh", "Updates for Support_Updater are mandatory. It'll only take a minute!\n\nSorry about that.");
+    return;
+  }
   %obj.glassSwatch.delete();
   %obj.delete();
   GlassUpdaterSupport::resize();
@@ -300,29 +305,23 @@ if($BLG::MM::UseUpdaterDefault $= "") {
 //updaterInterfaceDisplay(%refreshing) //Called when updates are ready to be displayed to the user.
 
 package GlassUpdaterSupportPackage {
-  function canvas::pushDialog(%this, %dlg) {
-    if(%dlg.getName() $= "updaterDlg" && !$BLG::MM::UseUpdaterDefault) {
-      echo("Pushing glass updater instead!");
-      GlassUpdaterSupport::pushGlassUpdater();
-      return;
-    } else {
-      return parent::pushDialog(%this, %dlg);
-    }
-  }
-
   function updaterInterfacePushItem(%item) {
-    GlassUpdaterSupport::pushItem(%item);
-    parent::updaterInterfacePushItem(%item);
+    if($BLG::MM::UseUpdaterDefault)
+      parent::updaterInterfacePushItem(%item);
+    else
+      GlassUpdaterSupport::pushItem(%item);
   }
 
   function updaterInterfaceDisplay(%refreshing) {
-    GlassUpdaterSupport::pushGlassUpdater(false);
-    parent::updaterInterfaceDisplay(%refreshing);
+    if($BLG::MM::UseUpdaterDefault)
+      parent::updaterInterfaceDisplay(%refreshing);
+    else
+      GlassUpdaterSupport::pushGlassUpdater(false);
   }
 
   function UpdaterDownloadTCP::setProgressBar(%this, %value) {
     if(!$BLG::MM::UseUpdaterDefault) {
-	     %queueObj = updater.queue.currentDownload;
+      %queueObj = updater.queue.currentDownload;
       GlassUpdaterSupport::updateProgressBar(%queueObj, %value);
     }
